+++ src/models/team.js
@@ -1,5 +1,6 @@
 import { lightColors } from 'Models/user';
-/* globals CDN_URL */
+import { }
+
 const cacheBuster = Math.floor(Math.random() * 1000);
 
 export const DEFAULT_TEAM_AVATAR = 'https://cdn.glitch.com/55f8497b-3334-43ca-851e-6c9780082244%2Fdefault-team-avatar.svg?1503510366819';
